Merging and create new Graphics Classes
Merging the new Graphics classes used to display the Game Board and the panel to display the Game Status.
The Cell class was edited in order to display a star in the case it is needed.
A class named Dice is added to the project.
The GameField class is renamed to GameBoard.
The Player class can now interact with the Game Board.